backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
fastapi[all]
ollama
